refactor: move hyperscaler apps
